+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Not yet released**
 
+- Added `TcpListener::from_std_poll` method for Windows.
 - Fixed accepting TCP sockets for 32-bit Windows failing out with "unsupported socket family" error ([GitHub issue](https://github.com/ferronweb/ferron/issues/662))
 
 ## `vibeio` 0.2.12